Moved notification manager, added error handling
